+++ tests/unit/nix/cases/net-vm-network.nix
@@ -40,6 +40,10 @@ let
       mtu = 1280;
       mssClamp = true;
       lan.allowEastWest = true;
+      homeLan.mdns = {
+        enable = true;
+        dnsmasqLocal.enable = true;
+      };
     };
     d2b.envs.safe = {
       lanSubnet = "10.30.0.0/24";
@@ -86,10 +90,13 @@ let
   workRuleset = workNet.networking.nftables.ruleset;
   safeRuleset = safeNet.networking.nftables.ruleset;
   obsRuleset = obsNet.networking.nftables.ruleset;
+  workDnsmasqServers = workNet.services.dnsmasq.settings.server;
 
   mssClampRule = "tcp flags syn tcp option maxseg size set rt mtu";
   lanToLanForwardRule = ''iifname "eth1" oifname "eth1" ct state new accept'';
   lanToUplinkAcceptRule = ''iifname "eth1" oifname "eth0" ct state new accept'';
+  mdnsHomeInputRule = ''iifname "home0" udp dport 5353 accept'';
+  mdnsLanInputRule = ''iifname "eth1" udp dport 5353 accept'';
 
   hasRule = ruleset: needle: lib.hasInfix needle ruleset;
 
@@ -250,6 +257,64 @@ in
   };
   "net-vm-network/safe-nft-lan-to-lan-forward-absent" = {
     expr = hasRule safeRuleset lanToLanForwardRule;
+    expected = false;
+  };
+
+  # ---- home-LAN mDNS is net-VM scoped and opt-in -----------------------
+  "net-vm-network/host-avahi-disabled" = {
+    expr = cfg.services.avahi.enable;
+    expected = false;
+  };
+  "net-vm-network/workload-avahi-disabled" = {
+    expr = workGuest.services.avahi.enable;
+    expected = false;
+  };
+  "net-vm-network/work-net-avahi-enabled" = {
+    expr = workNet.services.avahi.enable;
+    expected = true;
+  };
+  "net-vm-network/work-net-avahi-reflector-enabled" = {
+    expr = workNet.services.avahi.reflector;
+    expected = true;
+  };
+  "net-vm-network/work-net-avahi-open-firewall-disabled" = {
+    expr = workNet.services.avahi.openFirewall;
+    expected = false;
+  };
+  "net-vm-network/work-net-avahi-interfaces" = {
+    expr = workNet.services.avahi.allowInterfaces;
+    expected = [ "home0" "eth1" ];
+  };
+  "net-vm-network/safe-net-avahi-disabled" = {
+    expr = safeNet.services.avahi.enable;
+    expected = false;
+  };
+  "net-vm-network/work-net-mdns-home0-input-rule" = {
+    expr = hasRule workRuleset mdnsHomeInputRule;
+    expected = true;
+  };
+  "net-vm-network/work-net-mdns-eth1-input-rule" = {
+    expr = hasRule workRuleset mdnsLanInputRule;
+    expected = true;
+  };
+  "net-vm-network/safe-net-mdns-home0-input-rule-absent" = {
+    expr = hasRule safeRuleset mdnsHomeInputRule;
+    expected = false;
+  };
+  "net-vm-network/safe-net-mdns-eth1-input-rule-absent" = {
+    expr = hasRule safeRuleset mdnsLanInputRule;
+    expected = false;
+  };
+  "net-vm-network/work-net-dnsmasq-local-forward" = {
+    expr = builtins.elem "/local/127.0.0.1#53530" workDnsmasqServers;
+    expected = true;
+  };
+  "net-vm-network/work-net-mdns-local-resolver-service" = {
+    expr = builtins.hasAttr "d2b-mdns-local-resolver" workNet.systemd.services;
+    expected = true;
+  };
+  "net-vm-network/safe-net-mdns-local-resolver-service-absent" = {
+    expr = builtins.hasAttr "d2b-mdns-local-resolver" safeNet.systemd.services;
     expected = false;
   };